output: introduce BaseOutput.transfer
