+++ dvc/stage/imports.py
@@ -13,7 +13,7 @@ def _update_import_on_remote(stage, remote, jobs):
         )
 
     url = stage.deps[0].def_path
-    stage.outs[0].hash_info = stage.repo.cloud.transfer(
+    stage.outs[0].transfer(
         url, jobs=jobs, update=True, remote=remote, command="update"
     )
